fix margin analysis; update README to run in current env
